Created first version of Slate
Created the Slate object and wrote its script. The object has a Sign Sprite, that'll only show in the dark. When interacting with the region where the sign is, the Player will make the Slate Sprite come out of the floor. When interacting with the Slate, the Player will read a text.